refactor(storage): storage helper functions pulled out of component files and moved into their own file
